Change server cmd default listen port
