+++ server/server.go
@@ -73,6 +73,6 @@ func Run(cfg *Config) {
 	sr.PathPrefix("/{id:[0-9]+}").Handler(httpd.NoPrefix(http.FileServer(http.FS(fsys)))).Methods(http.MethodGet)
 	sr.PathPrefix("/").Handler(http.FileServer(http.FS(fsys))).Methods(http.MethodGet)
 
-	log.Printf("=== Listen Port: %s ===\n", cfg.Server)
-	log.Fatal(http.ListenAndServe(cfg.Server, sr))
+	log.Printf("=== Listen Port: %s ===\n", cfg.Http.Listen)
+	log.Fatal(http.ListenAndServe(cfg.Http.Listen, sr))
 }
